Refactored CalloutStyle and moved closeButton to the Callout-component
